requirements.txt
click==7.1.2
emoji==0.6.0
joblib==1.0.0
nltk==3.5
numpy==1.19.5
pandas==1.2.0
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
setuptools-git==1.2
six==1.15.0
tqdm==4.55.1
